Update clipboard, selection, util
